test: string split and char-at exception
